This is a code for the given assignment
The code involves three steps.

step 1: In this step, the given folder in which there are all the earlier .csv files are unzipped and read.
step2:  In this step, first of all, an earlier combined.csv file is removed. Then all the data from the remaining .csv files is extracted and combined. Then a new column is created which has rows assigned as basenames of the files from which the data is extracted. Then the duplicates are deleted and the values are sorted Then the file is saved as combined.csv
step3: In this step, the folder is zipped and given the same name as earlier.